chore: cleanup and improve the index
This change adds metrics to the various steps of the publication process on the index, adds support for categories and keywords, adds more outputs to the json for searches and crate listing, and adds a check to verify that something is in the index before downloading, fixing an issue where a crate pushed to the crates storage could potentially be visible even if the transaction failed to commit.

The download storage is always treated as safe to overwrite, so it is assumed that crates which may not be downloadable could be in there.