storage: fix bug in crate storage and separate paths
This fixes two bugs in the initial release. The first is that crates and indexes files would not be separated out into separate subdirectories, making bucket maintenance annoying.

The second, more serious one is that the use of the content encoding flag would cause the tarfiles to be unzipped on pull from the bucket, which is incorrect.