Improve client conformance to config.json
This change makes the client use auth when specified by the registry,
and obeys markers in URLs when present.

We still lack support for the sha256-checksum marker, but this requires
a public API change to support.

The e2e tests are also updated to conform to expected registry
behaviour.