RUST-77: implement list api (/all) for the FS-based index (#63)
This commit implements the /all endpoint for fs-based indexes.

We append to the end of an index entry the latest publish. That publish
data contains the latest relevant metadata for a crate.

The only thing that is missing is proper `created_at` and `updated_at`
dates. That can be added in a later PR.

We remove the `create_or_append` api. We simply use a ready-copy-update
approach instead.

A get to `/all` locally looks like this:
```
{
  "results": [
    {
      "name": "freighter-api-types",
      "versions": [
        {
          "version": "1.0.0"
        }
      ],
      "description": "Cloudflare's third-party Rust registry implementation",
      "created_at": "2024-08-08T12:43:35.293249Z",
      "updated_at": "2024-08-08T12:43:35.293250Z",
      "homepage": null,
      "repository": "https://github.com/cloudflare/freighter",
      "documentation": null,
      "keywords": [
        "registries",
        "freighter"
      ],
      "categories": []
    }
  ]
}
```